requirements_testing.txt
idna==2.9
black==19.10b0
flake8==3.8.2
mock==4.0.2
mypy==0.780
bandit==1.6.2
isort==4.3.21
safety==1.9.0
pylint==2.5.2
pytest-aiohttp==0.3.0
pytest==5.4.3
pytest-asyncio==0.12.0
pytest-cov==2.9.0
coverage==5.1
Sphinx==3.0.4
sphinx-rtd-theme==0.4.3
pip==20.1.1
wheel==0.34.2
tox==3.15.1
async-generator==1.10
async_timeout==3.0.1
docker-compose==1.26.0
watchdog==0.10.2
requests==2.23.0